requirements.txt
streamlit==1.52.1
torch==2.9.0
torchvision==0.24.0
diffusers==0.36.0
transformers==4.57.3
accelerate==1.12.0
Pillow==11.3.0
